admins shouldnt delete themselves

+++ app/Http/Controllers/ProfileController.php
@@ -42,7 +42,7 @@ class ProfileController extends Controller
     public function selfDestroyUser()
     {
         $user = Auth::user();
-        //if ($user->role == "admin") return back()->with("error", "You cannot delete yourself as an admin!");
+        if ($user->role == "admin") return back()->with("error", "You cannot delete yourself as an admin!");
 
         $user->delete();
